+++ eva/src/components/dashBoard.vue
@@ -74,7 +74,7 @@
           <div class="settings-dash">
             <v-dialog v-model="fullScreenMode">
               <template v-slot:activator="{ on: onFullScreen }">
-                <v-tooltip bottom :color="theme.$accent_ui_color">
+                <v-tooltip bottom :color="theme.$accent_ui_color" :open-delay="tooltipOpenDelay" :disabled="disabledTooltip">
                   <template v-slot:activator="{ on: onTooltip }">
                     <v-icon
                       class="expand"
@@ -152,7 +152,7 @@
                     </div>
                     <div class="settings-dash-block">
                       <div class="settings-dash">
-                        <v-tooltip bottom :color="theme.$accent_ui_color">
+                        <v-tooltip bottom :color="theme.$accent_ui_color" :open-delay="tooltipOpenDelay">
                           <template v-slot:activator="{ on }">
                             <v-icon
                               class="expand"
@@ -202,7 +202,7 @@
             :class="{ settings_move: props.open_gear }"
             v-show="dataMode"
           >
-            <v-tooltip bottom :color="theme.$accent_ui_color">
+            <v-tooltip bottom :color="theme.$accent_ui_color" :open-delay="tooltipOpenDelay">
               <template v-slot:activator="{ on }">
                 <v-icon
                   class="datasource"
@@ -215,7 +215,7 @@
               </template>
               <span>Источник данных</span>
             </v-tooltip>
-            <v-tooltip v-if="props.edit_icon" bottom :color="theme.$accent_ui_color">
+            <v-tooltip v-if="props.edit_icon" bottom :color="theme.$accent_ui_color" :open-delay="tooltipOpenDelay">
               <template v-slot:activator="{ on }">
                 <v-icon
                   class="pencil"
@@ -233,7 +233,7 @@
               </template>
               <span>Переименовать</span>
             </v-tooltip>
-            <v-tooltip v-if="!props.edit_icon" bottom :color="theme.$accent_ui_color">
+            <v-tooltip v-if="!props.edit_icon" bottom :color="theme.$accent_ui_color" :open-delay="tooltipOpenDelay">
               <template v-slot:activator="{ on }">
                 <v-icon
                   class="check"
@@ -246,7 +246,7 @@
               </template>
               <span>Переименовать</span>
             </v-tooltip>
-            <v-tooltip bottom :color="theme.$accent_ui_color">
+            <v-tooltip bottom :color="theme.$accent_ui_color" :disabled="disabledTooltip" :open-delay="tooltipOpenDelay">
               <template v-slot:activator="{ on }">
                 <v-icon
                   class="option"
@@ -259,7 +259,7 @@
               </template>
               <span>Настройки</span>
             </v-tooltip>
-            <v-tooltip bottom :color="theme.$accent_ui_color">
+            <v-tooltip bottom :color="theme.$accent_ui_color" :open-delay="tooltipOpenDelay">
               <template v-slot:activator="{ on }">
                 <v-icon
                   class="delete"
@@ -359,6 +359,10 @@ export default {
     },
     searchData: Array,
     dataSourseTitle: null,
+    tooltipOpenDelay: {
+      type: Number,
+      default: 500,
+    },
   },
   data() {
     return {
@@ -366,6 +370,7 @@ export default {
       mdiDatabaseSearch: mdiDatabaseSearch,
       mdiArrowDownBold: mdiArrowDownBold,
       fullScreenMode: false,
+      disabledTooltip: false,
       settings: {
         showTitle: true
       },
@@ -426,6 +431,9 @@ export default {
     };
   },
   computed: {
+    settingsIsOpened(){
+      return this.$store.getters.getModalSettings(this.idDash).status
+    },
     fullScreenWidth() {
       return 0.8 * window.innerWidth;
     },
@@ -530,6 +538,14 @@ export default {
       return options.change;
     },
     ...mapGetters(["getSelectedTableTitles", "getSelectedDataFormat"]),
+  },
+  watch: {
+    fullScreenMode(to) {
+      setTimeout(() => this.disabledTooltip = to, to ? 0 : 600)
+    },
+    settingsIsOpened(to){
+      setTimeout(() => this.disabledTooltip = to, to ? 0 : 600)
+    }
   },
   mounted() {
     this.props.icons = settings.icons;
@@ -1000,4 +1016,9 @@ export default {
 
 <style lang="scss">
 @import "../sass/dashBoard.sass";
+</style>
+<style lang="sass">
+  .settings-dash
+      .v-icon:focus::after
+          opacity: 0
 </style>